requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-jalali-date==1.0.2
idna==3.6
jdatetime==4.1.1
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
